JSON signed ints

all integers were interpreted as unsigned when rendered to JSON.

refs #9728

See merge request !20

+++ src/web/handler/propertyrequesthandler.cpp
@@ -107,7 +107,7 @@ namespace dss {
         if (node->getValueType() == vTypeInteger) {
           json.add("value", node->getIntegerValue());
         } else {
-          json.add("value", (long long int)node->getUnsignedIntegerValue());
+          json.add("value", node->getUnsignedIntegerValue());
         }
         return json.successJSON();
       } catch(PropertyTypeMismatch& ex) {
